Release Version 1.1
As of 10,5,23 or Wednesday the 10th of May 2023 there is a stable version of Simulacra. It is simplified, lacks scripts, random events, an interface and has plenty more work to be done. It does have a file system, upcoming deepening of detail as well as a working simulation of a small town, easy to understand and even easier to modify.

Specific Changes:
Reduced variable complexity, the way they were calculated was essentially random, so this was changed.

Reduced base economy values, these were set to that of a small town

Old files have been moved for brevity

Upcoming planned changes:

Citizen system: wealth, age, future wealth calculated for each individual

Country economy: hopefully resemble an economy of the UK to some degree of accuracy

Scripts: Events such as elections, pandemics should exist